separate authoenticate command from dto and result
