+++ src/Authorization/PL.Authorization/Authentication/AuthenticateCommand.cs
@@ -1,7 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Security.Claims;
-using System.Text;
+﻿using System.Text;
 using PL.BuildingBlocks.Application;
 
 namespace PL.Authorization.Application.Authentication
@@ -16,35 +13,5 @@ namespace PL.Authorization.Application.Authentication
             Login = login;
             Password = password;
         }
-    }
-
-    public class AuthenticationResult
-    {
-        public bool IsAuthenticated { get; }
-        public string AuthenticationError { get; }
-        public UserDto User { get; }
-
-        public AuthenticationResult(string authenticationError)
-        {
-            IsAuthenticated = false;
-            AuthenticationError = authenticationError;
-        }
-
-        public AuthenticationResult(UserDto user)
-        {
-            this.IsAuthenticated = true;
-            this.User = user;
-        }
-    }
-
-    public class UserDto
-    {
-        public Guid Id { get; set; }
-        public string Login { get; set; }
-        public bool IsActive { get; set; }
-        public string Name { get; set; }
-        public string Email { get; set; }
-        public List<Claim> Claims { get; set; }
-        public string Password { get; set; }
     }
 }
